hawawa
hawawa

+++ server/db.js
@@ -17,7 +17,7 @@ const db = {
     });
   },
   findUser (userInfo, cb) {//로그인 시
-    conn.query('select u_no, u_name from user where u_id = ? and u_pw = ?;',  [userInfo.uid, userInfo.upw], cb);
+    conn.query('select u_no, u_id, u_name from user where u_id = ? and u_pw = ?;',  [userInfo.uid, userInfo.upw], cb);
   },
   findUserByNo(u_no, cb) {//요청 시 토큰의 id값으로 유저 인증
     u_no = u_no * 1;
@@ -28,11 +28,14 @@ const db = {
       conn.query('update user set u_name = ?, u_email = ?, u_introduction = ? where u_no = ?', [u_name, u_email, u_intro, u_no], cb);
     }
   },
+  getLectureBySubscribed (uno, cb) {
+    conn.query('select l.l_no, l.l_wr_id, l.l_wr_name, l.l_title, l.l_thum, l.l_view, l.l_like from lecture l, subscribe s, user u where s.s_following_no = u.u_no and u.u_id = l.l_wr_id and s.s_follower_no = ?', uno, cb)
+  },
   getLectureWithNewest(cb) {
-    conn.query('select l_no, l_wr_id, l_wr_name, l_title, l_thum, l_view from lecture order by l_date desc limit 0, 5', cb);
+    conn.query('select l_no, l_wr_id, l_wr_name, l_title, l_thum, l_view, l_like from lecture order by l_date desc limit 0, 5', cb);
   },
   getLectureWithPopularity(cb) {
-    conn.query('select l_no, l_wr_id, l_wr_name, l_title, l_thum, l_view from lecture order by l_view desc limit 0, 5', cb);
+    conn.query('select l_no, l_wr_id, l_wr_name, l_title, l_thum, l_view, l_like from lecture order by l_view desc limit 0, 5', cb);
   },
   getNumberofLectures(uid, cb) {
     conn.query('select u_lectures from user where u_id = ?', uid, cb);
@@ -84,10 +87,10 @@ const db = {
     conn.query('insert into note(n_ts,n_string) value(?,?)',[timest,notestr]);
   },
   searchLectureByTitle(searchStr, cb) {
-    conn.query('select l_no, l_title, l_wr_id, l_wr_name, l_thum, l_view from lecture WHERE l_title REGEXP ?', searchStr, cb);
+    conn.query('select l_no, l_title, l_wr_id, l_wr_name, l_thum, l_view, l_like from lecture WHERE l_title REGEXP ?', searchStr, cb);
   },
   searchUserByName(searchStr, cb) {
-    conn.query('select u_id, u_email, u_introduction, u_lectures, u_follower from user WHERE u_id REGEXP ?' ,searchStr,cb);
+    conn.query('select u_id, u_name, u_email, u_introduction, u_lectures, u_follower from user WHERE u_name REGEXP ?' ,searchStr,cb);
   },
   searchLecturePartitioning(cb) {
     conn.query('select * from category where c_level1 = 00',cb);
@@ -101,12 +104,12 @@ const db = {
         conn.query('insert into like_history(u_no, l_idx) values(?, ?)', [uno, rows[0].l_idx], (err) => {
           if(!err){
             conn.query('update lecture set l_like = l_like + 1 where l_idx = ?', rows[0].l_idx)
-            cb(true)
+            cb('T')
           }
           else{
             conn.query('delete from like_history where u_no = ? and l_idx = ?', [uno, rows[0].l_idx], () => {
               conn.query('update lecture set l_like = l_like - 1 where l_idx = ?', rows[0].l_idx)
-              cb(false)
+              cb('F')
             })
           }
         })
@@ -117,21 +120,24 @@ const db = {
     conn.query('select u_no from user where u_id = ?', f_ing_id, (err, rows) => {
       let f_ing = rows[0].u_no
       //console.log(f_ing_id, rows)
-
-      conn.query('insert into subscribe(s_follower_no, s_following_no) values(?, ?)', [f_wer, f_ing], (err) => {
-        if(!err) {
-          conn.query('update user set u_following = u_following + 1 where u_no = ?', f_wer)
-          conn.query('update user set u_follower = u_follower + 1 where u_no = ?', f_ing)
-          cb(true)
-        }
-        else {
-          conn.query('delete from subscribe where s_follower_no = ? and s_following_no = ?', [f_wer, f_ing], (err) => {
-            conn.query('update user set u_following = u_following - 1 where u_no = ?', f_wer)
-            conn.query('update user set u_follower = u_follower - 1 where u_no = ?', f_ing)
-            cb(false)
-          })
-        }
-      })
+      if (f_ing === f_wer) {
+        cb('S')
+      } else {
+        conn.query('insert into subscribe(s_follower_no, s_following_no) values(?, ?)', [f_wer, f_ing], (err) => {
+          if(!err) {
+            conn.query('update user set u_following = u_following + 1 where u_no = ?', f_wer)
+            conn.query('update user set u_follower = u_follower + 1 where u_no = ?', f_ing)
+            cb('T')
+          }
+          else {
+            conn.query('delete from subscribe where s_follower_no = ? and s_following_no = ?', [f_wer, f_ing], (err) => {
+              conn.query('update user set u_following = u_following - 1 where u_no = ?', f_wer)
+              conn.query('update user set u_follower = u_follower - 1 where u_no = ?', f_ing)
+              cb('F')
+            })
+          }
+        })
+      }
     })
   }
 }
